[jit] split out refinements into their own file
I'm trying to reduce the overall complexity of the ir_emitter file and
this seemed like a reasonable candidate to split out and document
separately. Ideally we'd have tests for this logic as well

[ghstack-poisoned]
